moved tests inside component folders
