+++ .nvmrc
@@ -1,1 +1,1 @@
-  lts/carbon
+lts/carbon
